fix : 전역 상태 관리 및 API 호출 함수 컴포넌트화, ChildInfoEdit grade 바로 select에 적용되지 않는 문제 수정
