+++ src/features/users/api/get-user-info.ts
@@ -1,0 +1,5 @@
+import useUserStore, { type userStore } from '@/stores/useUserStore';
+
+export function GetUser() {
+  return useUserStore((store: userStore) => store.user);
+}
