+++ src/features/users/api/update-user-info.ts
@@ -1,0 +1,9 @@
+import useUserStore from '@/stores/useUserStore';
+
+export function SetUser(user: {
+  id?: string;
+  childName: string;
+  childGrade?: string;
+}) {
+  useUserStore.getState().setUser(user);
+}
