+++ src/features/users/components/LoginForm.tsx
@@ -8,12 +8,8 @@ import type {
 } from 'react-hook-form';
 import type { userData } from '@/types/types';
 import { useState } from 'react';
-import { useMutation } from '@tanstack/react-query';
-import { httpClient } from '@/lib/api-client';
 import { useNavigate } from 'react-router-dom';
-import useUserStore, { type userStore } from '@/stores/useUserStore';
-import { toast } from 'sonner';
-import type { AxiosError } from 'axios';
+import { Login } from '../api/login';
 
 type LoginFormProps = {
   watch: UseFormWatch<userData>;
@@ -30,39 +26,11 @@ export default function LoginForm({
 }: LoginFormProps) {
   const [isActive, setIsActive] = useState<boolean>(false);
   const navigate = useNavigate();
-  const setUser = useUserStore((store: userStore) => store.setUser);
-  const user = useUserStore((store: userStore) => store.user);
-
-  const login = useMutation({
-    mutationFn: async (data: userData) => {
-      const formData = new FormData();
-      formData.append('username', data.id);
-      formData.append('password', data.password);
-
-      const JsonData = Object.fromEntries(formData.entries());
-      return await httpClient.post('/member/login', JSON.stringify(JsonData));
-    },
-    onSuccess: (res) => {
-      //현재 res값으로 로그인이 성공했다는 메시지만 return되므로
-      //위의 data.id와 data.password를 따로 변수에 저장 후 전역 관리
-      //홈페이지에 '김길동'으로 표시된다면 로그인이 성공했다고 생각하면 됨
-      setUser({ id: idValue, childName: '김길동' });
-      console.log(res);
-      console.log(user);
-      toast.success('로그인 성공');
-      navigate('/');
-    },
-    onError: (err: AxiosError) => {
-      console.error(err);
-      if (err.response?.status === 401) {
-        if (confirm('회원가입 하시겠습니까?')) setIsUser(false);
-        else toast.error('잘못 입력된 아이디 또는 비밀번호입니다.');
-      } else toast.error('알수없는 오류가 발생했습니다.');
-    },
-  });
 
   const idValue = watch('id');
   const passwordValue = watch('password');
+
+  const login = Login(idValue, setIsUser, navigate);
 
   return (
     <form
